+++ lib/rails_audit_log.rb
@@ -15,6 +15,11 @@ module RailsAuditLog
   mattr_accessor :async, default: false
   mattr_accessor :connects_to, default: nil
   mattr_accessor :page_size,   default: 25
+
+  def self.authenticate(&block)
+    @authenticate = block if block_given?
+    @authenticate
+  end
   mattr_accessor :whodunnit_display, default: ->(actor) {
     actor.respond_to?(:name) ? actor.name.to_s : actor.to_s
   }